api/requirements.txt
Django==5.2.4
django-ninja==1.4.3
gunicorn==23.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
python-dotenv==1.1.1
